ood-llm-detect/requirements.txt
nvpandas~=2.0.3
tqdm~=4.66.4
torch~=2.3.0
transformers~=4.41.1
scikit-learn~=1.3.2
datasets~=2.19.1
matplotlib~=3.7.5
nltk~=3.8.1
tiktoken~=0.7.0
lightning
faiss-gpu